+++ examples/js/examples.js
@@ -1092,7 +1092,8 @@ NGL.ExampleRegistry.addDict( {
         stage.loadFile( "data://ala3.pdb" ).then( function( o ){
 
             var atomPair = [
-                [ "1.CA", "3.CA" ]
+                // [ "1.CA", "3.CA" ]
+                [ 8, 28 ]
             ];
 
             o.addRepresentation( "licorice" );
@@ -1226,7 +1227,7 @@ NGL.ExampleRegistry.addDict( {
     "bondOrders": function( stage ) {
 
         stage.loadFile( "data://4umt_47w.sdf" ).then( function ( o ) {
-            o.addRepresentation( "licorice", { multipleBond: true } );
+            o.addRepresentation( "licorice", { multipleBond: "symmetric" } );
             stage.centerView();
         } );
 
@@ -1235,7 +1236,7 @@ NGL.ExampleRegistry.addDict( {
     "chemCompCif": function( stage ){
 
         stage.loadFile( "data://PRDCC_000001.cif" ).then( function( o ){
-            o.addRepresentation( "licorice", { sele: "/0", multipleBond: true } );
+            o.addRepresentation( "licorice", { sele: "/0", multipleBond: "symmetric" } );
             stage.centerView();
         } );
 
@@ -1259,6 +1260,63 @@ NGL.ExampleRegistry.addDict( {
 
     },
 
+    "spatialHash": function( stage ) {
+
+        stage.loadFile( "rcsb://3sn6.mmtf", {
+            defaultRepresentation: false
+        } ).then( function( o ){
+
+            // o.addRepresentation( "backbone", { lineOnly: true } );
+            o.addRepresentation( "cartoon", { quality: "low" } );
+            stage.centerView();
+
+            var radius = 8;
+            var s = o.structure;
+            var spatialHash = new NGL.SpatialHash( s.atomStore, s.boundingBox );
+
+            var spacefillRepr = o.addRepresentation( "ball+stick", { sele: "NONE"/*, radius: 0.5*/ } );
+
+            function getCenterArray(){
+                var position = new NGL.Vector3();
+                var target = stage.viewer.controls.target;
+                var group = stage.viewer.rotationGroup.position;
+                position.copy( group ).negate().add( target );
+                return position;
+            }
+
+            function getSele( pos ){
+                var within = spatialHash.within( pos.x, pos.y, pos.z, radius );
+                return within.length ? "@" + within.join( "," ) : "NONE";
+            }
+
+            var sphereBuffer = new NGL.SphereBuffer(
+                new Float32Array( getCenterArray().toArray() ),
+                new Float32Array( [ 1, 0.5, 1 ] ),
+                new Float32Array( [ radius ] )
+            );
+            o.addBufferRepresentation( sphereBuffer, { opacity: 0.5 } );
+
+            var prevSele = "";
+            var prevPos = new NGL.Vector3( Infinity, Infinity, Infinity );
+            stage.viewer.controls.addEventListener(
+                'change', function(){
+                    var pos = getCenterArray();
+                    if( pos.distanceTo( prevPos ) > 0.1 ){
+                        sphereBuffer.setAttributes( { "position": pos.toArray() } );
+                        prevPos = pos;
+                        var sele = getSele( pos );
+                        if( sele !== prevSele ){
+                            spacefillRepr.setSelection( sele );
+                            prevSele = sele;
+                        }
+                    }
+                }
+            );
+
+        } );
+
+    },
+
     "axes": function( stage ){
 
         stage.loadFile( "rcsb://3pqr.mmtf", {
@@ -1278,6 +1336,31 @@ NGL.ExampleRegistry.addDict( {
             o.addRepresentation( "surface", { visible: false, lazy: true } );
             stage.centerView();
         } );
+
+    },
+
+    ringFlags: function( stage ) {
+
+        stage.loadFile( "rcsb://4w93.mmtf" ).then( function( o ){
+
+            o.addRepresentation( "licorice", { sele: "[3L9]" } );
+
+            // Get ring atoms for residue with name 3L9
+            var ringAtomNames = [];
+            o.structure.eachAtom( function( ap ) {
+                if( ap.isRing() ){
+                    ringAtomNames.push( "." + ap.atomname );
+                }
+            }, new NGL.Selection( "[3L9]" ));
+
+            o.addRepresentation("spacefill", {
+                sele: "[3L9] and ( " + ringAtomNames.join(" ") + ")",
+                scale: 0.25
+            });
+
+            stage.centerView();
+
+        })
 
     },
 
@@ -1408,7 +1491,7 @@ NGL.ExampleRegistry.addDict( {
                 cylinderOnly: true
             } );
             stage.centerView();
-        });
+        } );
 
     },
 
@@ -1469,6 +1552,6 @@ NGL.ExampleRegistry.addDict( {
             stage.centerView();
         });
 
-    },
+    }
 
 } );