+++ src/buffer/text-buffer.js
@@ -5,7 +5,7 @@
  */
 
 
-import { CanvasTexture } from "../../lib/three.es6.js";
+import { Color, CanvasTexture } from "../../lib/three.es6.js";
 
 import { Browser } from "../globals.js";
 import { defaults } from "../utils.js";
@@ -268,6 +268,17 @@ TextAtlas.prototype = {
  * @property {Float} xOffset - offset in x-direction
  * @property {Float} yOffset - offset in y-direction
  * @property {Float} zOffset - offset in z-direction (i.e. in camera direction)
+ * @property {String} attachment - attachment of the label, one of:
+ *                                 "bottom-left", "bottom-center", "bottom-right",
+ *                                 "middle-left", "middle-center", "middle-right",
+ *                                 "top-left", "top-center", "top-right"
+ * @property {Boolean} showBorder - show border/outline
+ * @property {Color} borderColor - color of the border/outline
+ * @property {Float} borderWidth - width of the border/outline
+ * @property {Boolean} showBackground - show background rectangle
+ * @property {Color} backgroundColor - color of the background
+ * @property {Float} backgroundMargin - width of the background
+ * @property {Float} backgroundOpacity - opacity of the background
  */
 
 
@@ -298,6 +309,14 @@ function TextBuffer( position, size, color, text, params ){
     this.xOffset = defaults( p.xOffset, 0.0 );
     this.yOffset = defaults( p.yOffset, 0.0 );
     this.zOffset = defaults( p.zOffset, 0.5 );
+    this.attachment = defaults( p.attachment, "bottom-left" );
+    this.showBorder = defaults( p.showBorder, false );
+    this.borderColor = defaults( p.borderColor, "lightgrey" );
+    this.borderWidth = defaults( p.borderWidth, 0.15 );
+    this.showBackground = defaults( p.showBackground, false );
+    this.backgroundColor = defaults( p.backgroundColor, "lightgrey" );
+    this.backgroundMargin = defaults( p.backgroundMargin, 0.5 );
+    this.backgroundOpacity = defaults( p.backgroundOpacity, 1.0 );
 
     var n = position.length / 3;
 
@@ -308,6 +327,7 @@ function TextBuffer( position, size, color, text, params ){
 
     this.text = text;
     this.count = charCount;
+    if( this.showBackground ) this.count += n;
     this.positionCount = n;
 
     this.vertexShader = "SDFFont.vert";
@@ -320,7 +340,12 @@ function TextBuffer( position, size, color, text, params ){
         "xOffset": { value: this.xOffset },
         "yOffset": { value: this.yOffset },
         "zOffset": { value: this.zOffset },
-        "ortho": { value: 0.0 }
+        "ortho": { value: false },
+        "showBorder": { value: this.showBorder },
+        "borderColor": { value: new Color( this.borderColor ) },
+        "borderWidth": { value: this.borderWidth },
+        "backgroundColor": { value: new Color( this.backgroundColor ) },
+        "backgroundOpacity": { value: this.backgroundOpacity }
     } );
 
     this.addAttributes( {
@@ -356,7 +381,12 @@ TextBuffer.prototype = Object.assign( Object.create(
         sdf: { updateShader: true, uniform: true },
         xOffset: { uniform: true },
         yOffset: { uniform: true },
-        zOffset: { uniform: true }
+        zOffset: { uniform: true },
+        showBorder: { uniform: true },
+        borderColor: { uniform: true },
+        borderWidth: { uniform: true },
+        backgroundColor: { uniform: true },
+        backgroundOpacity: { uniform: true }
 
     }, Buffer.prototype.parameters ),
 
@@ -418,13 +448,14 @@ TextBuffer.prototype = Object.assign( Object.create(
         var iCharAll = 0;
         var txt, iChar, nChar;
 
-        for( var v = 0; v < n; v++ ) {
+        for( var v = 0; v < n; ++v ) {
 
             o = 3 * v;
             txt = text[ v ];
             nChar = txt.length;
+            if( this.showBackground ) nChar += 1;
 
-            for( iChar = 0; iChar < nChar; iChar++, iCharAll++ ) {
+            for( iChar = 0; iChar < nChar; ++iChar, ++iCharAll ) {
 
                 for( var m = 0; m < 4; m++ ) {
 
@@ -478,40 +509,79 @@ TextBuffer.prototype = Object.assign( Object.create(
 
         var ta = this.textAtlas;
         var text = this.text;
+        var attachment = this.attachment;
+        var margin = ( ta.lineHeight * this.backgroundMargin * 0.1 ) - 10;
 
         var inputTexCoord = this.geometry.attributes.inputTexCoord.array;
         var inputMapping = this.geometry.attributes.mapping.array;
 
         var n = this.positionCount;
-
-        var c;
-        var i;
         var iCharAll = 0;
-        var txt, xadvance, iChar, nChar;
+        var c, i, txt, xadvance, iChar, nChar, xShift, yShift;
 
-        for( var v = 0; v < n; v++ ) {
+        for( var v = 0; v < n; ++v ) {
 
             txt = text[ v ];
             xadvance = 0;
             nChar = txt.length;
 
-            for( iChar = 0; iChar < nChar; iChar++, iCharAll++ ) {
+            // calculate width
+            for( iChar = 0; iChar < nChar; ++iChar ) {
+                c = ta.mapped[ txt[ iChar ] ];
+                xadvance += c.w - 2 * ta.outline;
+            }
+
+            // attachment
+            if( attachment.startsWith( "top" ) ){
+                yShift = ta.lineHeight / 1.25;
+            }else if( attachment.startsWith( "middle" ) ){
+                yShift = ta.lineHeight / 2.5;
+            }else{
+                yShift = 0;  // "bottom"
+            }
+            if( attachment.endsWith( "right" ) ){
+                xShift = xadvance;
+            }else if( attachment.endsWith( "center" ) ){
+                xShift = xadvance / 2;
+            }else{
+                xShift = 0;  // "left"
+            }
+            xShift += ta.outline;
+            yShift += ta.outline;
+
+            // background
+            if( this.showBackground ){
+                i = iCharAll * 2 * 4;
+                inputMapping[ i + 0 ] = -ta.lineHeight/6 - xShift - margin;  // top left
+                inputMapping[ i + 1 ] = ta.lineHeight - yShift + margin;
+                inputMapping[ i + 2 ] = -ta.lineHeight/6 - xShift - margin;  // bottom left
+                inputMapping[ i + 3 ] = 0 - yShift - margin;
+                inputMapping[ i + 4 ] = xadvance + ta.lineHeight/6 - xShift + 2*ta.outline + margin;  // top right
+                inputMapping[ i + 5 ] = ta.lineHeight - yShift + margin;
+                inputMapping[ i + 6 ] = xadvance + ta.lineHeight/6 - xShift + 2*ta.outline + margin;  // bottom right
+                inputMapping[ i + 7 ] = 0 - yShift - margin;
+                inputTexCoord[ i + 0 ] = 10;
+                inputTexCoord[ i + 2 ] = 10;
+                inputTexCoord[ i + 4 ] = 10;
+                inputTexCoord[ i + 6 ] = 10;
+                iCharAll += 1;
+            }
+
+            xadvance = 0;
+
+            for( iChar = 0; iChar < nChar; ++iChar, ++iCharAll ) {
 
                 c = ta.mapped[ txt[ iChar ] ];
                 i = iCharAll * 2 * 4;
 
-                // top left
-                inputMapping[ i + 0 ] = xadvance - ta.outline;
-                inputMapping[ i + 1 ] = c.h - ta.outline;
-                // bottom left
-                inputMapping[ i + 2 ] = xadvance - ta.outline;
-                inputMapping[ i + 3 ] = 0 - ta.outline;
-                // top right
-                inputMapping[ i + 4 ] = xadvance + c.w - ta.outline;
-                inputMapping[ i + 5 ] = c.h - ta.outline;
-                // bottom right
-                inputMapping[ i + 6 ] = xadvance + c.w - ta.outline;
-                inputMapping[ i + 7 ] = 0 - ta.outline;
+                inputMapping[ i + 0 ] = xadvance - xShift;  // top left
+                inputMapping[ i + 1 ] = c.h - yShift;
+                inputMapping[ i + 2 ] = xadvance - xShift;  // bottom left
+                inputMapping[ i + 3 ] = 0 - yShift;
+                inputMapping[ i + 4 ] = xadvance + c.w - xShift;  // top right
+                inputMapping[ i + 5 ] = c.h - yShift;
+                inputMapping[ i + 6 ] = xadvance + c.w - xShift;  // bottom right
+                inputMapping[ i + 7 ] = 0 - yShift;
 
                 var texWidth = ta.width;
                 var texHeight = ta.height;