+++ src/representation/line-representation.js
@@ -19,13 +19,13 @@ import LineBuffer from "../buffer/line-buffer.js";
  * @param {Structure} structure - the structure to be represented
  * @param {Viewer} viewer - a viewer object
  * @param {RepresentationParameters} params - representation parameters, plus the properties listed below
- * @param {Boolean} params.multipleBond - whether or not to render multiple bonds
+ * @property {String} multipleBond - one off "off", "symmetric", "offset"
  * @param {Float} params.bondSpacing - spacing for multiple bond rendering
  * @param {null} params.flatShaded - not available
  * @param {null} params.side - not available
  * @param {null} params.wireframe - not available
  * @param {null} params.roughness - not available
- * @param {null} params.matelness - not available
+ * @param {null} params.metalness - not available
  * @param {null} params.diffuse - not available
  */
 function LineRepresentation( structure, viewer, params ){
@@ -45,11 +45,17 @@ LineRepresentation.prototype = Object.assign( Object.create(
     parameters: Object.assign( {
 
         multipleBond: {
-            type: "boolean", rebuild: true
+            type: "select", rebuild: true,
+            options: {
+                "off" : "off",
+                "symmetric" : "symmetric",
+                "offset": "offset"
+            }
         },
         bondSpacing: {
-            type: "number", precision: 2, max: 1.0, min: 0.5
+            type: "number", precision: 2, max: 2.0, min: 0.5
         }
+
 
     }, Representation.prototype.parameters, {
 
@@ -67,8 +73,8 @@ LineRepresentation.prototype = Object.assign( Object.create(
 
         var p = params || {};
 
-        this.multipleBond = defaults( p.multipleBond, false );
-        this.bondSpacing = defaults( p.bondSpacing, 0.85 );
+        this.multipleBond = defaults( p.multipleBond, "off" );
+        this.bondSpacing = defaults( p.bondSpacing, 1.0 );
 
         StructureRepresentation.prototype.init.call( this, p );
 
@@ -130,9 +136,7 @@ LineRepresentation.prototype = Object.assign( Object.create(
         var what = {};
 
         if( params && params.bondSpacing ){
-
             what.position = true;
-
         }
 
         StructureRepresentation.prototype.setParameters.call(