+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file, following t
 - chainname based color scheme
 - BondHash class to quickly get atoms connected to an atom
 - XmlParser parameter to use the browser's DOMParser
+- attachment (top, middle, bottom; left, center, right) for LabelRepresentation/TextBuffer
+- border color and width for LabelRepresentation/TextBuffer
+- colored background rectangle for LabelRepresentation/TextBuffer
+- "offset" style rendering of double/triple bonds (@fredludlow)
 
 ### Changed
 - ResidueindexColorMaker colorscale domain on a per chain basis